Google Cloud Platform/Data-Orchestration-Cloud-Function/2. prospectEstimator/requirements.txt
functions-framework==3.*
google-cloud-bigquery
pandas
db-dtypes
fsspec
gcsfs
pandas-gbq
flask
google-cloud-pubsub
python-liquid
numpy
google-auth